+++ java/org/apache/catalina/startup/CredentialHandlerRuleSet.java
@@ -30,7 +30,7 @@ public class CredentialHandlerRuleSet implements RuleSet {
 
     private static final int MAX_NESTED_LEVELS = Integer.getInteger(
             "org.apache.catalina.startup.CredentialHandlerRuleSet.MAX_NESTED_LEVELS",
-            3).intValue();
+            3);
 
     // ----------------------------------------------------- Instance Variables
