+++ java/org/apache/coyote/http2/AbstractStream.java
@@ -50,7 +50,7 @@ abstract class AbstractStream {
 
 
     final int getIdAsInt() {
-        return identifier.intValue();
+        return identifier;
     }
 
 
@@ -124,11 +124,11 @@ abstract class AbstractStream {
         if (windowSize > ConnectionSettingsBase.MAX_WINDOW_SIZE) {
             String msg = sm.getString("abstractStream.windowSizeTooBig", getConnectionId(), identifier,
                     Integer.toString(increment), Long.toString(windowSize));
-            if (identifier.intValue() == 0) {
+            if (identifier == 0) {
                 throw new ConnectionException(msg, Http2Error.FLOW_CONTROL_ERROR);
             } else {
                 throw new StreamException(
-                        msg, Http2Error.FLOW_CONTROL_ERROR, identifier.intValue());
+                        msg, Http2Error.FLOW_CONTROL_ERROR, identifier);
             }
         }
     }
