+++ java/org/apache/tomcat/util/log/UserDataHelper.java
@@ -70,7 +70,7 @@ public class UserDataHelper {
         // Default suppression time of 1 day.
         suppressionTime = Integer.getInteger(
                 "org.apache.juli.logging.UserDataHelper.SUPPRESSION_TIME",
-                60 * 60 * 24).intValue() * 1000L;
+                60 * 60 * 24) * 1000L;
 
         if (suppressionTime == 0) {
             tempConfig = Config.INFO_ALL;
