+++ java/org/apache/tomcat/util/net/SocketWrapperBase.java
@@ -1030,10 +1030,10 @@ public abstract class SocketWrapperBase<E> {
     protected class VectoredIOCompletionHandler<A> implements CompletionHandler<Long, OperationState<A>> {
         @Override
         public void completed(Long nBytes, OperationState<A> state) {
-            if (nBytes.longValue() < 0) {
+            if (nBytes < 0) {
                 failed(new EOFException(), state);
             } else {
-                state.nBytes += nBytes.longValue();
+                state.nBytes += nBytes;
                 CompletionState currentState = state.isInline() ? CompletionState.INLINE : CompletionState.DONE;
                 boolean complete = true;
                 boolean completion = true;
